rename markdown-clean to markdown, sync cache-control headers, add edge injection
