+++ www/src/pages/llms.txt.ts
@@ -27,7 +27,7 @@ ${links}
   return new Response(body, {
     headers: {
       "Content-Type": "text/plain; charset=utf-8",
-      "Cache-Control": "public, max-age=3600",
+      "Cache-Control": "public,max-age=0,s-maxage=86400,stale-while-revalidate=86400",
     },
   });
 };
